mtd: spi-nor: Cadence QSPI driver needs to enable its clock.
On the Arria10 SoC, the PL330 DMAC and Cadence QSPI share a clock source.
It was discovered that when the PL330 and QSPI were both enabled in the
device tree, the QSPI would timeout.  We traced this to the AMBA probe
function, which leaves the DMAC clock disabled.  This also disabled the
derived QSPI clock.  Even if the PL330 driver was subsequently loaded,
the derived QPSI clock was not re-enabled.

Regardless of that issue, the devm_get_clk documentation specifically
states that "Drivers must assume that the clock source is not enabled."
Which means the QSPI driver should enable its clock.  This patch also
disables the clock on driver removal.

Signed-off-by: Graham Moore <grmoore@opensource.altera.com>