HSD #1707131464: EDAC/altera: Add S10 SDRAM ECC check for legacy U-Boot
A bug in U-Boot causes a crash during SDRAM boot if ECC is
not enabled in the bitstream but is enabled in the Linux
config.[1]

Memory mapped read of the ECC Enabled bit was only enabled
if U-Boot determined ECC was enabled in the bitstream.

The Linux driver checks the ECC enable bit using a memory map
read. In the ECC disabled bitstream case, U-Boot didn't
enable ECC register memory map reads and since they are not
allowed this results in a crash.

This patch always reads the ECC Enable register through an
SMC call which is always allowed and will work with legacy
and current U-Boot.

[1] https://lkml.org/lkml/2019/11/29/366

Fixes: Commit 08f08bfb7b4c ("EDAC, altera: Merge Stratix10 into the Arria10 SDRAM probe routine")

Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>
(cherry picked from commit 0dfe1db9a44d611e54bede6bbd77bcb342d65eb4)
Signed-off-by: Grygorii Tertychnyi <grygorii.tertychnyi@leica-geosystems.com>