Remove offset pointer optimization in inftrees.c.
commit faadf9935688f65618dc84e42248ffdb78796cf0 from U-Boot
Addresses CVE-2016-9840

inftrees.c was subtracting an offset from a pointer to an array,
in order to provide a pointer that allowed indexing starting at
the offset. This is not compliant with the C standard, for which
the behavior of a pointer decremented before its allocated memory
is undefined. Per the recommendation of a security audit of the
zlib code by Trail of Bits and TrustInSoft, in support of the
Mozilla Foundation, this tiny optimization was removed, in order
to avoid the possibility of undefined behavior.

Signed-off-by: Mark Adler <madler@alumni.caltech.edu>
Signed-off-by: Chin Liang See <chin.liang.see@intel.com>
Signed-off-by: Thor Thayer <thor.thayer@linux.intel.com>
(cherry picked from commit 648a7c9c7698ac162427270fa3bf1ef3ce00023d)
Signed-off-by: Grygorii Tertychnyi <grygorii.tertychnyi@leica-geosystems.com>