arm64: dts: ap20-pt1: set USB hub to default config
As I2C3 lines are interchanged, no communication with USB2514B is
possible. CPU_USBH_I2C_SDA (instead of CPU_USBH_I2C_SCL) is connected to
pin `SCL/CFG_SEL0` of USB2514B and set to HIGH by a pullup resistor on
SOM. Therefore the initial interface configuration of USB2514B is set to
SMBus and a configuration over I2C3 is expected.
In order to load the default configuration of USB2514B the `CFG_SEL[0]`
needs to be set to 0 i.e. hog GPIO5_IO19 to low before releasing reset.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>