arm: krea_defconfig: disable building with gcc plugins
GCC plugins are broken in Yocto SDK, as they are disabled in SDK GCC,
but Kernel build system checks them on the host, which does provide
them. This leads to a false identification that plugins are enabled and
leads to further build faliures.

Disable building KREA kernel with GCC plugins, as they are unusable
until fixed in Yocto SDK.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>