mtd: spi-nor: add altera quadspi controller driver
Introduce support for Intel FPGA Generic QUAD SPI Controller Core
Driver, which is used to communicate with FPGA Configuration Flash
devices and program FPGA fabric.

This driver covers only Version 1 of the Controller, which can be
instantiated in FPGA fabric. Note that the Controller implemented
in FGPA is not JEDEC-compliant, therefore not all opcodes are supported
by this driver.

Driver is based on Altera QuadSPI Controller driver, which is not
present in upstream, and has been originally proposed in [1].

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>
Link: [1]: https://lwn.net/Articles/655154/