ARM: dts: leica-krea: correct properties for quadspi
QuadSPI node has a missing reg property in the child node, which is
pasred by the driver as a chip select. Also, compatible string set for
child node does not match to what in set for configuration devices.

Add missing property and set configuration device chip select to 0
according to HW layout.

Correct compatible string to point to correct match in the driver.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>