spi: dw: send messages in atomic context with irq disabled
Some of the slaves (e.g. EDM Monitor) connected to this controller
misbehave when messages are not sent in a contiguous sequence
(i.e. when slave select signal is de-asserted during tranmission,
resulting in a message being transmitted in two or more parts).

Send SPI messages in an atomic context.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>