crypto: caam - bump down CAAM hashing cra_priority
Tests showed that ARMv8-A crypto extensions as implemented on the ARM
Cortex-A53 on the i.MX8M family of processors are substantially faster
than the same primitives offered by the CAAM. The 3000 CRA priority
specified for the CAAM since its upstreaming puts it so far ahead that
it will always be chosen by the kernel if available. Bump it down a few
notches to 150, which will allow the 200 cra_priority of the crypto
extensions to be chosen instead if available.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>