arm64: dts: crocodile: decrease I2C clocks
According to Mask Set Errata IMX8MM_0N87W, the clock frequency must be
set to 384kHz to fulfil the requirement of a low period of 1.3us.

Decrease clock frequency of I2C3 to 100kHz (since we do not have timing
critical communications on this interface) and set it to the suggested
value of 384kHz for I2C4.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>