HSD #1308655004: don't clear the ue/ce status bits
When there is a uncorrected or double-bit error, the kernel will panic
and will result in a watchdog reset. When the SDM gets the reset, it
needs to know that a DBE had occurred so it can scrub the DDR on a
reboot. Thus we should not clear the CE/UE status bits in the EDAC.

Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>
(cherry picked from commit 43263c93f47096774fabb4e7024aeef9730ea8b6)
Signed-off-by: Grygorii Tertychnyi <grygorii.tertychnyi@leica-geosystems.com>