arm64: dts: crocodile: remove DMA for console
Operations on kernel console needs to be atomic - that's why `uart2`
in imx8mm.dtsi does not provide any DMA-properties, since usually this
UART is used as kernel console.
On crocodile `uart3` is used as console.

Remove DMA-properties from node `uart3` and add the specific ones to
`uart2` (which are missing in imx8mm.dtsi).

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>