iio: st_sensors: add alwayson flag
The st_sensors_read_info_raw() implementation allows to get raw data
from st_sensors, enabling and disabling the device at every read.
This leads to delays in data access, because, after every enable, we
need to wait the hardware to be ready.

Introducing the 'alwayson' flag in (st_sensor_data) and the
correspondent sysfs attribute, we allow the user to keep the device
always enabled. In this way, every data access to the device can be
performed with no delays.

Signed-off-by: Massimo Toscanelli <massimo.toscanelli@leica-geosystems.com>