HSD #14015950233: crypto: intel_fcs: select HW_RANDOM if intel_fcs is built-in
Update Kconfig to add the dependencies of intel_fcs module. This change
will only build in the HW_RANDOM when the FCS is enabled. So if FCS is a
kernel module, then HW_RANDOM is a hw module, if FCS is built-in, then the
HW_RANDOM is built-in.

Fixes: 6153ec6 ("HSD #22013011814: crypto: intel_fcs: support HWRNG via /dev/hwrng")

Signed-off-by: Teh Wen Ping <wen.ping.teh@intel.com>
(cherry picked from commit c7c2b484981cacc53579786da50ce3edcd9adfea)
Signed-off-by: Grygorii Tertychnyi <grygorii.tertychnyi@leica-geosystems.com>