PCI: add class quirk for Microchip DG0756 to fix BAR assignment
The Microchip DG0756 FPGA image uses a class code of 0
(PCI_CLASS_NOT_DEFINED). This prevents the PCI core from assigning
resources for the BARs because __dev_sort_resources() ignores
classless devices, host bridges, and IOAPICs.

On x86, firmware typically assigns those resources, so this was not a
problem.  But on some architectures, firmware does *not* assign BARs, and
since the PCI core didn't do it either, the DG0576 device didn't work
correctly.

Add the quirk to the PCI core, where it will always run early enough that
the PCI core will assign resources if necessary.

Signed-off-by: Matthias Fend <matthias.fend@emfend.at>
Signed-off-by: Elfendi <elfendi@emfend.at>