soc: imx: imx8mp-blk-ctrl: set the LCDIF hurry priority
Status: WIP, needs clarification with NXP first.

Set the LCDIF hurry priority to highest possible value so FIFO underruns
can be avoided. The hurry priority will be set by the BLKCTL hw as soon
as the LCDIF panic signal is active and set back to 'normal' priority
after the panic signal is released.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>