arm64: dts: crocodile: uart2: set IMX8MM_SYS_PLL1_80M parent of uart2 clock
The UART baudrate in the Bluetooth firmware of uBlox BT/WLAN module
MAYA-W161 on PT1 is set to 3M (while it is 115200 on MAYA-W160).

Without setting IMX8MM_SYS_PLL1_80M as parent, the baudrate of 3M is too
high for the default clock input and the communication with the module
fails.
As baudrate 115200 for MAYA-W160 can also be achieved with
IMX8MM_SYS_PLL1_80M this goes into `crocodile.dtsi`.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>