arm64: dts: crocodile: align device tree naming with schematics
We support three versions of crocodile board revisions
- Rev. B:
  On this revision, there is no separation of GPIO modules between M4
  and A53.
  On this board, we don't consider any M4 access to the modules (e.g.
  ecspi3 is enabled).
- Rev. B reworked:
  Reworked boards, where we have dedicated access from M4 and A53 to
  GPIO modules.
  Due to the rework, IO-multiplexer values of some unused GPIOs needs to
  be set.
- Rev. C:
  Boards, where we have dedicated access from M4 and A53 to GPIO
  modules.

The file names of the device trees, as well as the model node, are
aligned with the revision names of the boards.

To support testing of the M4 dedicated modules from A53, we provide the
device tree overlay crocodile-no-m4.dts for reworked Rev. B and Rev. C
boards.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>