arm64: beluga_defconfig: fix imx-sdma fw load
If the imx-sdma driver is built in the Kernel, the SDMA HW is probed
when the firmware is not yet available.
So, we build the imx-sdma driver as a kernel module.

This has been fixed upstream in
e95622289f263662240544a9f0009b25c19e64d4 ("arm64: defconfig: build imx-sdma as a module")

Signed-off-by: Massimo Toscanelli <massimo.toscanelli@leica-geosystems.com>