arm64: crocodile_defconfig: enable powerkey handling
for the init-system to initiate a proper shutdown when the physical
power button is pressed briefly both the SNVS_PWRKEY driver as well as
the whole EVDEV interface need to be enabled.

the first so that the interrupt passed into the ONOFF pin of the imx8
is translated into the KEY_POWER event.  and the second that the
init-system "gets" the keyboard event and is able to initiate the
system shutdown

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>