net: phy: add phy_device_atomic_register helper
Currently the usually way to probe/setup a phy is done by:
 1) get_phy_device()/phy_device_create()
 2) phy_device_register.

During get_phy_device() the PHYID1/2 registers are read and the code
relies on the fact that the phy is accesibble. This is not the case if
the pre-running software like the bootloader did not initialize the phy.
Even if the bootloader initialized it, during the probe clocks can
disappear and so the phy is not accessible anymore.

To fix this we need to parse the phy's fwnode first, then do some basic
setup like: bring it out of the reset state and finally read the
PHYID1/2 registers.

Since the mdio/phy handling has a few legacy paths and current systems
should not break, this patch adds a new helper. The helper bundles the
two steps into one. So in the end: allocate, parse, init and register a
phy is done atomically. Since firmware parsing is required this commit
also copies the phy firmware node parsing helpers from fwnode_mdio.c to
this file. So the phy-parsing is bundled within the phy_device.c file
and not somewhere else. After we converterd all callers of
fwnode_mdiobus_* to this new API we can remove the support from
fwnode_mdio.c.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>