net: phy: add phy_device_atomic_register helper
Currently the usually way to probe and setup a phy is done via:
 1) get_phy_device()/phy_device_create()
 2) phy_device_register.

During get_phy_device() the PHYID1/2 registers are read which assumes
that the phy is already accessible. This is not always the case, e.g.
 - if the pre-running firmware did not initialize the phy or
 - if the kernel does gate important clocks while booting and the phy
   isn't accessible after the pre-running firmware anymore.

To fix this we need to:
 - parse the phy's fwnode first,
 - do some basic setup like: bring it out of the reset state and
 - finally read the PHYID1/2 registers to probe the correct driver

This patch adds a new helper called phy_device_atomic_register() to not
break exisiting running systems based on the current mdio/phy handling.
This new helper bundles all required steps into a single function to
make it easier for driver developers.

To bundle the phy firmware parsing step within phx_device.c the commit
copies the required code from fwnode_mdio.c. After we converterd all
callers of fwnode_mdiobus_* to this new API we can remove the support
from fwnode_mdio.c.

Upstream-Status: Pending

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>