net: mdio: make use of phy_device_atomic_register helper
The current fwnode_mdiobus_register_phy() implementation assume that the
phy is accessible to read the PHYID register values. This assumption is
wrong since the firmware can left the phy in reset state or the phy
hardware dependencies like the phy-clock(s) are turned on/off during
booting which can confuse the phy internal state machine. Fix this by
using the new phy_device_atomic_register() helper which will ensure that
the prerequisites are fulfilled before accessing the PHXID registers.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>