net: mdio: make use of phy_device_atomic_register helper
The current fwnode_mdiobus_register_phy() implementation assume that the
phy is accessible to read the PHYID register values first which isn't
the case in some cases. Fix this by using the new
phy_device_atomic_register() helper which ensures that the prerequisites
are fulfilled before accessing the PHYID registers.

Upstream-Status: Pending

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>