net: mdiobus: remove now unused fwnode helpers
These APIs are broken since the very first day because they assume that
the phy is accessible to read the PHYID registers. If this requirement
is not meet the code fails to add the required phys.

The newly added phy_device_atomic_register() API have fixed this and
supports firmware parsing as well. After we switched all in kernel users
of the fwnode API we now can remove this part from the kernel.

Upstream-Status: Pending

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>