net: mdiobus: remove now unused fwnode helpers
These APIs are broken since the very first day since they assumed that
the phy is accessible to read the PHYID registers. If this requirement
is not meet the code fails to add the required phys.

To fix this we added the phy_device_atomic_register() API which is the
new atomic way to parse, setup and register a phy device. After we
switched all users of the fwnode API we now can remove this part from
the kernel.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>