net: phy: add default gpio assert/deassert delay
There are phy's not mention any assert/deassert delay within their
datasheets but the real world showed that this is not true. They need at
least a few us to be accessible and to readout the register values. So
add a sane default value of 1000us for both assert and deassert to fix
this in a global matter.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>