arm64: beluga_defconfig: add missing interconnect option
i.MX8M+ SOC has it's own config option for interconnect, which is not
selected in the Beluga config and also in defconfig from upstream.

linux-next have the upstream commit 0cca034d0234 ("arm64: defconfig:
enable i.mx 8m plus specific interconnect support") to fix the same, but
it sets the interconnect option to be built as module. We do set this as
built-in for our Kernel.

Add missing interconnect option and sync the config via `savedefconfig`
target.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>