media: rkisp1: tpg: Add support for custom frame sizes and rates
Add support to the TPG module for custom frame sizes and custom frame
rates.

The validation for frame rates is not strict enough yet and some
selections (mainly edge cases) will result in no frames at one extreme
or missed irqs at the other extreme, but the average/common frame size
and frame rate this will work fine.

This is for the TPG module that is on the version of the ISP that is on
the i.MX8MP.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
Reviewed-by: Umang Jain <umang.jain@ideasonboard.com>
Tested-by: Daniel Scally <dan.scally@ideasonboard.com>

---
Changes in v2:
- initialize the interval so that it can run at a reasonable default
  value even when an fps isn't supplied