media: rkisp1: Support setting crop at runtime
Add support for setting the crop at runtime. Previously the crop could
only be set at configure time, and could not be changed at runtime. The
scaler component already worked properly, so this in effect enables
scaler crop.

Note that this is specifically for the i.MX8mp. As the rk3399 version
supports dual crop, it doesn't actually need to rely on the image
stabilizer for cropping. The rk3399 should still work with this, though
(needs testing), but it needs to be taken into account when support for
actual image stabilization is added later.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>