net: phy: add phy_id_broken support
Some phy's don't report the correct phy-id, e.g. the TJA1102 dual-port
report 0 for the 2nd port. To fix this a driver needs to supply the
phyid instead and tell the phy framework to not try to readout the
phyid. The latter case is done via the new 'phy_id_broken' flag which
tells the phy framework to skip phyid readout for the corresponding phy.

Upstream-Status: Pending

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>