arm64: beluga_defconfig: enable dma-buf CMA heap
Enable the dma-buf heap backed by the Contiguous Memory Allocator (CMA).
This enables the allocation of physically contiguous memory from user
space.
For beluga this allows a very efficient use of the DMA IP cores integrated
in the FPGA. It is possible to allocate and directly map dma capable
memory in the application software. Since the driver for the FPGA supports
the import of such buffers, this area can be filled directly from the IP
core. No further copying is necessary.

Signed-off-by: Matthias Fend <matthias.fend@emfend.at>