media: rkisp1: Update resizer shadow registers automatically
The ISP (both the version in the rk3399 and the one in the i.MX8MP)
supports automatically updating the shadow registers of all subsequent
modules in the pipeline (resizer, memory interface, etc) when the last
pixel in a frame is finished. This alleviates the need for manually
timing when to update the shadow registers.

This feature was previously unused, causing an image size mismatch and
thus a pipeline stall when zooming in while doing digital zoom, where
the crop rectangle (in the image stabilizer, in the ISP) was changed
/before/ the scaling (in the resizer) was changed. That is, the image
size was shrunk by the ISP while the resizer still expected the old
bigger image size.

Fix this issue by changing the shadow register update mechanism to
auto-update when the ISP is done processing the frame.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>