media: rkisp1: Fix desync of crop and scaler in digital zoom
When implementing digital zoom via scaling and cropping, there was an
issue where the crop would be updated before scaler. This causes a
problem where, when zooming in (ie. the crop rectangle gets smaller), the
crop rectangle is shrunk already but the resizer still expects the older
bigger rectangle.

This problem was partially fixed before [1] by using the auto-update
configuration feature of the ISP, where it automatically updates all
shadow registers of the following blocks when the ISP is done processing
a frame.

This was insufficient however, as the *non-shadow* registers for
cropping and scaling could still be programmed at different timings, and
thus there was a potential window when the ISP could update the shadow
registers after the crop non-shadow registers were programmed, but
before the scaler non-shadow registers were programmed.

Fix this issue by spinlocking programming of the crop and scaler
non-shadow registers with respect to the ISP updating the shadow
registers.

To achieve this, we cache the crop rectangles of the ISP output and the
resizer input subdevices, and decouple computing the register values and
programming the registers. The former has the added benefit of
supporting Rockchip's versions of the ISP, which have a dual crop at the
output of the ISP (which NXP's don't), as well as paving the way to
support future NXP's versions of the ISP which potentially have a
resizer crop at the input of the resizer.

[1] "media: rkisp1: Update resizer shadow registers automatically"

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>