media: rkisp1: resizer: Clamp crop to valid values
The resizer (at least the one on the i.MX8MP) is unable to upscale to
higher than 1080p, or higher than x2 (whichever is lower), and if it is
set the stream will stall. Prevent this by adding the scaler restriction
to rkisp1_rsz_set_sink_crop().

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>
---
This limit seems to be not strict enough.