arm64: dts: freescale: imx8mp: rename imx8mp-isp[12].dtso
Rename imx8mp-isp[12].dtso to imx8mp[01].dtso. The only place where 1
based indexes are used is the reference manual. In all other places
the numbering is 0 based. Let's stick to that to avoid confusion. It
also maps better to dt-overlays referencing a specific csi port.
E.g. imx8mp-isp0.dtbo + imx8mp-some-sensor-on-csi0.dtbo

This is already included in the v6.9 tree:
https://gitlab.com/ideasonboard/nxp/linux/-/tree/v6.9/dt/imx8mp?ref_type=heads

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>