wifi: mwifiex: increase max_num_akm_suites
The maximum number of AKM suites will be set to two if not specified by
the driver. Set it to CFG80211_MAX_NUM_AKM_SUITES to let userspace
specify up to ten AKM suites in the akm_suites array.

Without only the first two AKM suites will be used, further ones are
ignored.

Acked-by: Brian Norris <briannorris@chromium.org>
Link: https://lore.kernel.org/r/20240530130156.1651174-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>