media: imx283: Add CENTERED_RECTANGLE helper
This reverts "media: imx283: drop CENTERED_RECTANGLE due to clang
failure" which is applied upstream. The CENTERED_RECTANGLE helper was
dropped from mainline as it's not really used there, and fails to build
on clang, but we do want to make use of it for our other custom modes.

Upstream-Status: Pending

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>